refactors and messing with logic in test files.
refactors to draw the cursor correctly and fixing logic bugs,
but there needs to be some forced organization here with regards
to when/how things are drawn.

need to think about the specific parameters that functions need to take
and what they mean; like should a function accept the prior cursor position
or recalculate based on the rope?

rope data structure management and usage.

tests and more.