refactor: Unifies all folding logic for all GADT variants in rope.ml and put types in their own file to avoid cycles
Puts folding logic for Rope_Traversal_Info, Line_Numbers, and Finding_Cursor into rope.ml along with traverse_rope.

Changes the traverse_rope function so that the interface allows a function to be passed in with the accumulated info and
glyph and process it and return unit. Allows for less duplication of logic when traversing the rope and makes the
interface less fragile.